Cleaning up file structure, removing redundant docstrings
