+++ tests/test_cinos.py
@@ -1,65 +1,48 @@
-from api.cinos import Drink, Order
-
-""" Drink class tests section """
+from api.Drink import Drink
+from api.Order import Order
 
 def test_get_base():
-    """
-    Test the getter for the base of the drink.
-    """
+    """Test the getter for the base of the drink."""
     drink = Drink("sbrite")
     assert drink.get_base() == "sbrite"
 
 def test_get_flavors_empty():
-    """
-    Test case to check if get_flavors returns an empty list when flavors are not set.
-    """
+    """Test case to check if get_flavors returns an empty list when flavors are not set."""
     drink = Drink("water")
     assert drink.get_flavors() == []
 
 def test_get_total():
-    """
-    Test if get_total method returns the correct drink price.
-    """
+    """Test if get_total method returns the correct drink price."""
     drink = Drink("water")
     assert drink.get_total() == 1
 
 def test_get_num_flavors_if_empty():
-    """
-    Test if get_num_flavors returns 0 when no flavors are set.
-    """
+    """Test if get_num_flavors returns 0 when no flavors are set."""
     drink = Drink("water")
     assert drink.get_num_flavors() == 0
 
 def test_get_num_flavors_if_not_empty():
-    """
-    Test if get_num_flavors returns the correct number of flavors when flavors are set.
-    """
+    """Test if get_num_flavors returns the correct number of flavors when flavors are set."""
     drink = Drink("water")
     flavors = ["lemon", "mint", "blueberry"]
     drink.set_flavors(flavors)
     assert drink.get_num_flavors() == len(flavors)
 
 def test_set_flavors():
-    """
-    Test if set_flavors method correctly sets the flavors.
-    """
+    """Test if set_flavors method correctly sets the flavors."""
     drink = Drink("water")
     flavors = ["lemon", "mint", "blueberry"]
     drink.set_flavors(flavors)
     assert drink.get_flavors() == flavors
 
 def test_add_flavor():
-    """
-    Test adding a valid flavor using the add_flavor method.
-    """
+    """Test adding a valid flavor using the add_flavor method."""
     drink = Drink("water")
     drink.add_flavor("lemon")
     assert drink.get_flavors() == ["lemon"]
 
 def test_duplicate_add_flavor():
-    """
-    Test adding a duplicate flavor and check if the correct exception is raised.
-    """
+    """Test adding a duplicate flavor and check if the correct exception is raised."""
     drink = Drink("water")
     drink.add_flavor("lemon")
     try:
@@ -67,42 +50,30 @@ def test_duplicate_add_flavor():
     except ValueError as ve:
         assert str(ve) == "Invalid or duplicate flavor"
 
-""" Order class test section """
-
 def test_get_items_when_empty():
-    """
-    Test if get_items returns an empty array when no items are added to order
-    """
+    """Test if get_items returns an empty array when no items are added to the order."""
     order = Order()
     assert order.get_items() == []
 
-def test_get_num_items_when_emtpy():
-    """
-    Test if get_num_items returns 0 when no items are added to the order.
-    """
+def test_get_num_items_when_empty():
+    """Test if get_num_items returns 0 when no items are added to the order."""
     order = Order()
     assert order.get_num_items() == 0
 
 def test_get_total_when_empty():
-    """
-    Test if get_total returns 0 when no items are added to the order.
-    """
+    """Test if get_total returns 0 when no items are added to the order."""
     order = Order()
     assert order.get_total() == 0
 
 def test_add_item():
-    """
-    Test if add_item correctly adds a drink to the order.
-    """
+    """Test if add_item correctly adds a drink to the order."""
     order = Order()
     drink = Drink("water")
     order.add_item(drink)
     assert order.get_items() == [drink]
 
 def test_remove_item_valid_index():
-    """
-    Test if remove_item removes the item at a valid index.
-    """
+    """Test if remove_item removes the item at a valid index."""
     order = Order()
     drink1 = Drink("water")
     drink2 = Drink("sbrite")
@@ -112,9 +83,7 @@ def test_remove_item_valid_index():
     assert order.get_items() == [drink2]
 
 def test_remove_item_invalid_index():
-    """
-    Test if remove_item does nothing for an invalid index.
-    """
+    """Test if remove_item does nothing for an invalid index."""
     order = Order()
     drink = Drink("water")
     order.add_item(drink)
